Server: Split Model.cs, rename PalServerContext
