+++ Pal.Server/Database/GlobalSetting.cs
@@ -1,0 +1,19 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Pal.Server
+{
+    public class GlobalSetting
+    {
+        public GlobalSetting(string key, string value)
+        {
+            Key = key;
+            Value = value;
+        }
+
+        [Key]
+        public string Key { get; set; }
+
+        [MaxLength(128)]
+        public string Value { get; set; }
+    }
+}
